feat: abstract data source
